# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERP API",
        "description": "List of all api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/store": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store user",
                "operationId": "efad1bddb09c97653ade562ab1dfb950",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "middle_name",
                                    "last_name",
                                    "dob",
                                    "joining_date",
                                    "permanent_province",
                                    "permanent_district",
                                    "permanent_municipality",
                                    "permanent_tole",
                                    "current_province",
                                    "current_district",
                                    "current_municipality",
                                    "current_tole",
                                    "department_id",
                                    "position_id",
                                    "role_id",
                                    "email",
                                    "staff_id",
                                    "mobile_number_1",
                                    "mobile_number_2",
                                    "profile_image",
                                    "citizenship_front",
                                    "citizenship_back"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "John"
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Cena"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2080/01/09"
                                    },
                                    "joining_date": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2080/01/09"
                                    },
                                    "permanent_province": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "permanent_district": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2"
                                    },
                                    "permanent_municipality": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "7"
                                    },
                                    "permanent_tole": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Laligurans"
                                    },
                                    "current_province": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2"
                                    },
                                    "current_district": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "4"
                                    },
                                    "current_municipality": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "7"
                                    },
                                    "current_tole": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Ganesthan"
                                    },
                                    "department_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2"
                                    },
                                    "position_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "4"
                                    },
                                    "role_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "2"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "john@gmail.com"
                                    },
                                    "staff_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "5"
                                    },
                                    "mobile_number_1": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "9829345674"
                                    },
                                    "mobile_number_2": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "9829345674"
                                    },
                                    "profile_image": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "profile.jpg"
                                    },
                                    "citizenship_front": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "citizenship_front.jpg"
                                    },
                                    "citizenship_back": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "citizenship_back.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully user saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/show": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List all users",
                "operationId": "5b2f8d2790414ef59a6903edac64a1da",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "sno": {
                                                        "type": "numeric",
                                                        "example": "1"
                                                    },
                                                    "id": {
                                                        "type": "numeric",
                                                        "example": "4"
                                                    },
                                                    "full_name": {
                                                        "type": "string",
                                                        "example": "John Cena"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@gmail.com"
                                                    },
                                                    "is_active": {
                                                        "type": "string",
                                                        "example": "Y"
                                                    },
                                                    "dob": {
                                                        "type": "string",
                                                        "example": "2080/01/02"
                                                    },
                                                    "staff_id": {
                                                        "type": "string",
                                                        "example": "234"
                                                    },
                                                    "joining_date": {
                                                        "type": "string",
                                                        "example": "2080/01/02"
                                                    },
                                                    "profile_image": {
                                                        "type": "string",
                                                        "example": "profile.jpg"
                                                    },
                                                    "primary_mobile_number": {
                                                        "type": "string",
                                                        "example": "9823453464"
                                                    },
                                                    "department_name": {
                                                        "type": "string",
                                                        "example": "IT"
                                                    },
                                                    "role_name": {
                                                        "type": "string",
                                                        "example": "Employee"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/user/delete/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "operationId": "3933578fb342e084f66164fe0c222608",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the User",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "25"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully user deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}